Fix number formatting errors for accurate display
Correct ambiguous number formatting calls and resolve missing type reference in number conversion logic.

+++ Converters/EnglishNumberConverter.cs
@@ -15,7 +15,7 @@ public class EnglishNumberConverter : IValueConverter
             decimal decimalValue => NumberFormatting.Format(decimalValue, format),
             double doubleValue => NumberFormatting.Format(doubleValue, format),
             float floatValue => NumberFormatting.Format((decimal)floatValue, format),
-            int intValue => NumberFormatting.Format(intValue, format),
+            int intValue => NumberFormatting.Format((decimal)intValue, format),
             _ => value?.ToString() ?? string.Empty
         };
     }
